Refactor - reorganized classes into corresponding packages
