+++ src/main/java/com/productmanager/entity/ProductManager.java
@@ -1,5 +1,12 @@
 package com.productmanager.entity;
 
+import com.productmanager.entity.product.Drink;
+import com.productmanager.entity.product.Food;
+import com.productmanager.entity.product.Product;
+import com.productmanager.entity.review.Rateable;
+import com.productmanager.entity.review.Rating;
+import com.productmanager.entity.review.Review;
+
 import java.math.BigDecimal;
 import java.text.MessageFormat;
 import java.text.NumberFormat;
