+++ src/main/java/com/productmanager/entity/product/Food.java
@@ -1,4 +1,6 @@
-package com.productmanager.entity;
+package com.productmanager.entity.product;
+
+import com.productmanager.entity.review.Rating;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
@@ -9,11 +11,11 @@ import java.time.LocalDate;
  * such as best before.
  */
 
-public class Food extends Product{
+public class Food extends Product {
 
     private LocalDate bestBefore;
 
-    Food(int id, String name, BigDecimal price, Rating rating, LocalDate bestBefore) {
+    public Food(int id, String name, BigDecimal price, Rating rating, LocalDate bestBefore) {
         super(id, name, price, rating);
         this.bestBefore = bestBefore;
     }
