+++ src/main/java/com/productmanager/entity/review/Rateable.java
@@ -1,4 +1,4 @@
-package com.productmanager.entity;
+package com.productmanager.entity.review;
 
 @FunctionalInterface
 public interface Rateable <T> {
