+++ src/main/java/com/productmanager/entity/review/Review.java
@@ -1,4 +1,6 @@
-package com.productmanager.entity;
+package com.productmanager.entity.review;
+
+import com.productmanager.entity.review.Rating;
 
 public class Review {
     private Rating rating;
